\name{GeneExpExample5000}
\alias{GeneExpExample5000}
\title{GeneExpExample5000}
\description{
  GeneExpExample5000.txt includes the first 5000 lines in SupplementaryTable2.txt which is a 
  supplementary file for Marioni,J.C. et al. (2008) (\url{http://genome.cshlp.org/content/18/9/1509/suppl/DC1}). 
}
\references{
  Marioni,J.C. et al. (2008) RNA-seq: an assessment of technical reproducibility and comparison with
  gene expression arrays. \emph{Genome Res.}, \bold{18}, 1509-1517.
}
\seealso{
 \code{\link{DEGexp}},
 \code{\link{getGeneExp}},
 \code{\link{readGeneExp}},
 \code{\link{samWrapper}},
 \code{\link{GeneExpExample1000}}.
}
\keyword{datasets}
