\name{kidneyChr21Bowtie}
\alias{kidneyChr21Bowtie}
\title{kidneyChr21Bowtie}
\description{
  The reads uniquely mapped to human chromosome 21 obtained from the kidney sample sequenced in Run 1, Lane 1.
}
\references{
  Marioni,J.C. et al. (2008) RNA-seq: an assessment of technical reproducibility and comparison with
  gene expression arrays. \emph{Genome Res.}, \bold{18}, 1509-1517.
}
\seealso{
 \code{\link{DEGexp}},
 \code{\link{DEGseq}},
 \code{\link{getGeneExp}},
 \code{\link{readGeneExp}},
 \code{\link{liverChr21.bed}},
 \code{\link{refFlatChr21}}.
}
\keyword{datasets}
