\name{refFlatChr21}
\alias{refFlatChr21}
\title{refFlatChr21}
\description{
  The gene annotation file includes the annotations of genes on chromosome 21, and is in UCSC refFlat format.
  See \url{http://genome.ucsc.edu/goldenPath/gbdDescriptionsOld.html#RefFlat}.
}
\seealso{
 \code{\link{DEGseq}},
 \code{\link{DEGexp}},
 \code{\link{kidneyChr21.bed}},
 \code{\link{liverChr21.bed}}.
}
\keyword{datasets}
